+++ bundled/minimal/src/minimal_clj/user.clj
@@ -1,0 +1,1 @@
+(ns minimal-clj.user)