+++ src/joyride.ts
@@ -46,8 +46,8 @@ export function isJoyrideNReplServerRunning() {
   }
 }
 
-export async function prepareForJackingOrConnect() {
-  await state.initProjectDir(ConnectType.JackIn, undefined).catch((e) => {
+export async function prepareForJackInOrConnect() {
+  await state.initProjectDir(ConnectType.JackIn, undefined, false, true).catch((e) => {
     void vscode.window.showErrorMessage('Failed initializing project root directory: ', e);
   });
   inspector.revealOnConnect();